Multi Agent System With MCP Server/sk_multi_agent_system/pyproject.toml
[project]
name = "sk-multi-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.7.1",
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.29.0",
]
